Move RelatedProductsSection to shared product directory, remove cart wrapper
Now lives at components/product/ since it's used by both PDP and cart.
