+++ apps/template/components/product/related-products-section.tsx
